+++ NEWS.md
@@ -22,6 +22,8 @@
 
 * `search_texts()` to search for words or glosses in the texts.
 
+* `show_entry()` which prints the entry with the given `entry_id`.
+
 ## Changed
 
 * `create_entry()` is now exported and outputs a list with the entry id and a string containing the entry skeleton.